+++ learning_observer/learning_observer/module_loader.py
@@ -22,7 +22,7 @@ import learning_observer.paths
 import learning_observer.settings
 
 from learning_observer.log_event import debug_log
-import learning_observer.communication_protocol.executor
+import learning_observer.communication_protocol.integration
 import learning_observer.queries
 import learning_observer.stream_analytics.helpers as helpers
 
@@ -307,7 +307,7 @@ def load_named_queries(component_name, module):
         debug_log(f"Loading named queries from {component_name}")
         # set lo.queries namespace
         queries = learning_observer.queries.NestedQuery()
-        learning_observer.communication_protocol.executor.add_queries_to_module(module.NAMED_QUERIES, queries)
+        learning_observer.communication_protocol.integration.add_queries_to_module(module.NAMED_QUERIES, queries)
         setattr(learning_observer.queries, component_name, queries)
 
         # clean queries
